lib/arcaea_prober_wrapper/requirements.txt
azure-functions
websockets
brotli

# Test
# pytest
